fix(common): resource rank table and members table optimization
