+++ shell/app/common/components/members-table/index.tsx
@@ -15,13 +15,13 @@ import { MemberScope } from 'common/stores/member-scope';
 import { usePerm } from 'app/user/common';
 import userStore from 'app/user/stores';
 import appMemberStore from 'common/stores/application-member';
-import { AddMemberModal, Copy, FilterGroup, FormModal, IF, DropdownSelect } from 'common';
+import { AddMemberModal, Copy, FilterGroup, FormModal, IF } from 'common';
 import { ColumnProps } from 'core/common/interface';
 import { useLoading } from 'core/stores/loading';
 import AuthorizeMemberModal from '../authorize-member-modal';
 import i18n from 'i18n';
 import { debounce, map, isEmpty, find, isArray, filter, get } from 'lodash';
-import { Button, Modal, Select, Spin, Tooltip, message } from 'antd';
+import { Button, Modal, Select, Spin, Tooltip, message, Avatar } from 'antd';
 import Table, { IActions } from 'common/components/table';
 import orgMemberStore from 'common/stores/org-member';
 import projectMemberStore from 'common/stores/project-member';
@@ -30,7 +30,7 @@ import React from 'react';
 import { useEffectOnce } from 'react-use';
 import UrlInviteModal from './url-invite-modal';
 import BatchAuthorizeMemberModal from '../batch-authorize-member-modal';
-import { insertWhen, goTo } from '../../utils';
+import { insertWhen, goTo, getAvatarChars } from 'common/utils';
 import { useUpdate } from '../../use-hooks';
 import routeInfoStore from 'core/stores/route';
 import memberLabelStore from 'common/stores/member-label';
@@ -388,7 +388,12 @@ const MembersTable = ({
                   (removed && i18n.t('exit the organization')) || ''
                 }`}
               >
-                <span>{nick || i18n.t('common:none')}</span>
+                <span>
+                  <Avatar src={record.avatar} size="small" className="mr-1">
+                    {getAvatarChars(nick || i18n.t('none'))}
+                  </Avatar>
+                  {nick || i18n.t('common:none')}
+                </span>
                 <IF check={currentUserId === userId}>
                   <span className="member-username-info"> [{i18n.t('current user')}]</span>
                 </IF>
