+++ shell/app/modules/dcos/pages/cluster-dashboard/resources-summary.scss
@@ -7,3 +7,7 @@
     color: $color-dark-8;
   }
 }
+
+.quota-usage-progress .ant-progress-circle-path {
+  transition: all 0.8s !important;
+}
