+++ shell/app/modules/dcos/pages/cluster-dashboard/resources-summary.tsx
@@ -195,10 +195,15 @@ const PureResourceTable = React.memo(({ rankType }: { rankType: string }) => {
   const { cpuPerNode, memPerNode } = cpuAndMem.current;
 
   const [data, loading] = getResourceTable.useState();
+  const [showPercent, setShowPercent] = React.useState(false);
   React.useEffect(() => {
     if (clusters.length) {
       const curCluster = clusterName?.length ? clusterName : clusters.map((item) => item.name);
-      getResourceTable.fetch({ clusterName: curCluster, cpuPerNode, memPerNode, groupBy: rankType });
+      getResourceTable.fetch({ clusterName: curCluster, cpuPerNode, memPerNode, groupBy: rankType }).then(() => {
+        setTimeout(() => {
+          setShowPercent(true); // To animate the progress bar
+        }, 100);
+      });
     }
   }, [clusters, clusterName, rankType, cpuPerNode, memPerNode]);
 
@@ -295,7 +300,8 @@ const PureResourceTable = React.memo(({ rankType }: { rankType: string }) => {
           <Tooltip title={`${record.cpuRequest} / ${record.cpuQuota}`}>
             <span className="text-dark-8  mr-2">{`${value}%`}</span>
             <Progress
-              percent={value}
+              className="quota-usage-progress"
+              percent={showPercent ? value : 0}
               type="circle"
               width={20}
               strokeWidth={18}
@@ -333,8 +339,9 @@ const PureResourceTable = React.memo(({ rankType }: { rankType: string }) => {
           <Tooltip title={`${record.memRequest} / ${record.memQuota}`}>
             <span className="text-dark-8 mr-2">{`${value}%`}</span>
             <Progress
+              className="quota-usage-progress"
               type="circle"
-              percent={value}
+              percent={showPercent ? value : 0}
               width={20}
               strokeWidth={18}
               format={(v) => null}
